Fix the way facts are set in the test suite
We must set the facts in the context of a specific operating system,
otherwise all operating systems will share the same facts.

This is currently not a problem because only Debian and Ubuntu are
supported and tested against, and Ubuntu being a Debian derivative both
OS facts are interchangeable, but adding support for more operating
systems trigger errors (e.g. FreeBSD).